requirements.txt
Django==1.5.5
Scrapy==0.20.0
Twisted==13.2.0
cssselect==0.9.1
django-tastypie==0.10.0
lxml==3.2.4
mimeparse==0.1.3
pyOpenSSL==0.13.1
python-dateutil==2.2
queuelib==1.1.1
six==1.4.1
w3lib==1.5
wsgiref==0.1.2
zope.interface==4.0.5
